Break out the dataset_conversion pieces
